evdev-lid: fix the notifier initial state
What we do not want is libinput to believe the LID is closed while
it's not.
But the internal notifier state need to be in sync with the evdev
node, or it's going to be a pain setting the keyboard listener.

We just can not forward the initial LIBINPUT_SWITCH_LID event when
registering the switch but we can store the internal state.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>